Moved the Alert extension into the iOS part
